Make benchmark crate usable as a library
